login logout and register

+++ foot/views.py
@@ -1,7 +1,10 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .form import RoomForm
 from django.contrib import messages
@@ -16,8 +19,14 @@ from django.contrib import messages
 
 def loginPage(request):
 
+    page = 'login'
+    # the below condition is very specifc one to handle direct url access to loginpage!!!
+    if request.user.is_authenticated:
+        return redirect('home')
+
+
     if request.method == "POST":
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -34,9 +43,28 @@ def loginPage(request):
         else:
             messages.error(request, 'username or password does not exist')
 
-    context= {}
+    context= {'page':page}
     return render(request, 'foot/login_register.html', context)
 
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occured during registration')
+
+    return render(request, 'foot/login_register.html', {'form':form})
 
 def Home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -55,8 +83,11 @@ def room(request, pk):
 # created a methos to request form from room_form.html
 # then go to foot/urls.py
 
+@login_required(login_url= 'login') #maing sure we let user to create the room if the user is logedin
 def createRoom(request):
     form = RoomForm()
+
+    # making sure of the form submission
     if request.method == 'POST':
         form= RoomForm(request.POST)
         if form.is_valid():
@@ -68,10 +99,16 @@ def createRoom(request):
     return render(request, 'foot/room_form.html', context)
 
 
+@login_required(login_url= 'login') #maing sure we let user to update the room if the user is logedin
 def updateRoom(request, pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance=room)
 
+# making sure usesr is only allowed tp update if it is logedin/Authaurized 
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
+
+# making sure of the form submission
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -81,9 +118,15 @@ def updateRoom(request, pk):
     context = {'form':form}
     return render(request, 'foot/room_form.html', context)
 
-
+@login_required(login_url= 'login')#maing sure we let user to delete the room if the user is logedin
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    # making sure user is only allowed to delete if it is logedin/Authaurized 
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
+    
+    # making sure of the form submission
     if request.method == 'POST':
         room.delete()
         return redirect('home')
